--- src/NoteEtudiant.java
@@ -1,33 +1,0 @@
-//ANCIEN SYSTEME
-//A SUPPRIMER
-class NoteEtudiant {
-
-    private int codeEtu;
-    private Double noteEtu;
-    private String classementEtu;
-
-    public int getCodeEtu() {
-        return codeEtu;
-    }
-
-    public void setCodeEtu(int codeEtu) {
-        this.codeEtu = codeEtu;
-    }
-
-    public Double getNoteEtu() {
-        return noteEtu;
-    }
-
-    public void setNoteEtu(Double valeur) {
-        this.noteEtu = valeur;
-    }
-    
-    public String getClassementEtu() {
-        return classementEtu;
-    }
-
-    public void setClassementEtu(String classementEtu) {
-        this.classementEtu = classementEtu;
-    }   
-
-}
